backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
twilio==9.4.0
python-jose[cryptography]==3.3.0
numpy==2.2.1
pydantic==2.10.4
python-multipart==0.0.18
bcrypt==4.1.2
PyJWT==2.12.1
slowapi==0.1.9
bleach==6.1.0
openai==1.3.9
openpyxl==3.1.5
python-docx==0.8.11
